Split Classes into seperate files, cleaned up create_item method
