+++ database_func.py
@@ -1,0 +1,11 @@
+import unittest
+from databases import Purchase, Lookup
+
+
+class Lookup_database_tests(unittest.TestCase):
+    def test_querying_categories(self):
+        self.assertEqual(True, True)  # add assertion here
+
+
+if __name__ == '__main__':
+    unittest.main()
