+++ lookup.csv
@@ -2,4 +2,6 @@ Name,Median_price,Category
 Chleb,5.99,Jedzenie
 Ser,19.99,Jedzenie
 Poduszki,15.59,Dekoracje
-Poduszka,12.59,Dekoracja
+Poduszka,12.59,Dekoracjachlyb
+@#!
+-1.0
